Fix after review

+++ Model/src/main/java/com/kurtsevich/rental/dto/rental_point/RentalPointWithDistanceDto.java
@@ -11,5 +11,5 @@ public class RentalPointWithDistanceDto {
     private int streetAddressNumber;
     private char streetAddressNumberSuffix;
     private String phoneNumber;
-    private int distance;
+    private long distance;
 }
